requirements.txt
# Geometry OS - Core Dependencies
# =================================

# Core scientific computing
numpy>=1.24.0
Pillow>=10.0.0

# Testing framework (core)
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-xdist>=3.0.0
pytest-asyncio>=0.23.0
pytest-timeout>=2.1.0

# Testing framework (extended)
pytest-html>=4.0.0
pytest-json-report>=1.5.0
pytest-randomly>=3.15.0
pytest-clarity>=1.0.0
pytest-benchmark>=4.0.0
hypothesis>=6.92.0

# Code quality
ruff>=0.1.0
mypy>=1.0.0

# Async/Networking
aiohttp>=3.8.0
aiofiles>=23.0.0
requests>=2.28.0

# Data processing
pyyaml>=6.0
toml>=0.10.2

# Utilities
psutil>=5.9.0

# Reed-Solomon error correction (for pixel compiler)
reedsolo>=1.7.0

# Template engine (for dashboard)
jinja2>=3.1.0
